z8000: fix arg_t to 32-bit for segmented pointers
arg_t was int16_t, truncating 32-bit segmented pointers passed as
syscall arguments. This caused _execve to receive a garbage pointer
for the /init path (0xFFFF8376 instead of 0x82008396).

Changed arg_t/uarg_t to int32_t/uint32_t to match other 32-bit FUZIX
ports (68000, riscv32). Updated kernel-z8000.def offsets for the
enlarged u_argn fields in struct u_data.

Current status: kernel mounts root fs, finds /init in directory, but
i_open returns wrong inode number (52 instead of 2). The directory
data is correct on disk. Root cause still under investigation —
likely a namei/directory lookup issue with the buffer cache or
struct layout.

+++ Kernel/cpu-z8000/types.h
@@ -11,8 +11,8 @@ typedef signed char int_fast8_t;
 
 typedef uint16_t irqflags_t;
 
-typedef int16_t arg_t;
-typedef uint16_t uarg_t;		/* Holds arguments */
+typedef int32_t arg_t;			/* Must hold pointers (32-bit segmented) */
+typedef uint32_t uarg_t;		/* Holds arguments */
 typedef uint16_t usize_t;		/* Largest value passed by userspace */
 typedef int16_t susize_t;
 typedef int16_t ssize_t;
